add weighted sampling
I weight each cluster, determined by its size relative to total number of images. and determine sample size per cluster, which allowed proportional representation. So larger the cluster, more samples are taken from the cluster. Moreover, this also is added with the centroid implementation. Therefore, it will sample the closest to centroid samples first.

So for example, if we have 3 clusters with sizes [100, 50, 25] and the user wants 35 samples:
Weights would be approximately [ 0.57, 0.29, 0.14]
Initial sample sizes: [20, 10,5]
After adjustments (ensuring minimums and not exceeding total): [20, 10,5]